requirements.txt
# requirements
pandas
numpy
statsmodels.api
pickle
matplotlib

